+++ CHANGELOG.md
@@ -5,5 +5,16 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
-## v0.1.0
+## [0.1.1] - 05.05.2018
+
+### Fixed
+- Use a unreleased version of `bip_metainfo` to fix the setter bugs
+
+### Changed
+- Non Verbose mode now shows a progress text of the hashed pieces.
+
+### Added
+- Multiple Announce URLs are now supported
+
+## [0.1.0]
 - Initial Version